% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_features_abd.R
\name{check_features_abd}
\alias{check_features_abd}
\title{check_features_abd}
\usage{
check_features_abd(feature_abd, comp)
}
\arguments{
\item{feature_abd}{a feature-by-sample data frame.}

\item{comp}{logical. If FALSE, the abundances are assumed non-compositional.}
}
\value{
error messages if feature abundances are (1) negative, and converts to proportions if comp is TRUE yet the abundances are not proportions.
}
\description{
function to ensure that feature abundances are neither missing nor negative. Missing abundances are replaced by zeros.
}
