% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultObjects.R
\name{control_adjust_batch}
\alias{control_adjust_batch}
\title{Default control settings for batch adjustment}
\usage{
control_adjust_batch
}
\description{
A named list of default parameters used in batch adjustment.
}
