% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultObjects.R
\name{control_lm_meta}
\alias{control_lm_meta}
\title{Default control settings for meta-analysis}
\usage{
control_lm_meta
}
\description{
A named list of default parameters used in the meta-analysis step.
}
