% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractScore.R
\name{extractScore}
\alias{extractScore}
\title{extractScore}
\usage{
extractScore(
  x,
  cl,
  comp,
  net.est.method,
  net_ctrl,
  verbose = TRUE,
  return_net = TRUE
)
}
\arguments{
\item{x}{a sample-by-feature abundance matrix.}

\item{cl}{cluster environment for resampled network estimation in parallel.}

\item{comp}{logical; TRUE for compositional data.}

\item{net.est.method}{network estimation method specified by the user according to the (non)compositionality of abundances.}

\item{net_ctrl}{named list of control parameters specific to the network estimation method specified.}

\item{verbose}{logical. If TRUE, print out number of iterations and computational time. Default value: TRUE.}

\item{return_net}{logical. If TRUE, the estimated network based on all the subjects is returned as part of the output. Default value: TRUE.}
}
\value{
a list with
\item{thetahat}{marginal network centrality measures based on all the subjects.}
\item{thetahat_loo}{resampled marginal network centrality measures leaving-one-subject at a time.}
\item{pseudoVal}{sample-by-feature matrix of pseudovalues.}
\item{estimatedNet}{the estimated network based on all the subjects; returned only when \code{returm_net} is TRUE.}
}
\description{
function to compute sample-by-feature pseudovalue matrix from sample-by-feature abundance matrix.
}
