% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_abdfeatures.R
\name{filter_abdfeatures}
\alias{filter_abdfeatures}
\title{filter_abdfeatures}
\usage{
filter_abdfeatures(x, abd_threshold, prev_threshold)
}
\arguments{
\item{x}{a sample-by-feature data frame.}

\item{abd_threshold}{threshold for abundance of features.}

\item{prev_threshold}{threshold for prevalence of features.}
}
\value{
a sample-by-feature data frame.
}
\description{
function to remove features according to user-specified abundance and prevalence threshold values.
}
