% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_features.R
\name{filter_features}
\alias{filter_features}
\title{filter_features}
\usage{
filter_features(x, abd_threshold, prev_threshold)
}
\arguments{
\item{x}{a sample-by-feature data frame.}

\item{abd_threshold}{threshold for abundance of features.}

\item{prev_threshold}{threshold for prevalence of features.}
}
\value{
sample-by-feature data frame after removing features.
}
\description{
function to remove features as per provided abundance and prevalence threshold values,
and zero variance features.
}
