% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_varfeatures.R
\name{filter_varfeatures}
\alias{filter_varfeatures}
\title{filter_varfeatures}
\usage{
filter_varfeatures(x, topV = NULL)
}
\arguments{
\item{x}{a sample-by-feature data frame.}

\item{topV}{number of most variable features to extract. Default is same as the number of features.}
}
\value{
a sample-by-feature data frame.
}
\description{
function to remove features with zero variance and keep a fixed number of the most variable features from the rest.
}
