% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measureNetwork.R
\name{measureNetwork}
\alias{measureNetwork}
\title{measureNetwork}
\usage{
measureNetwork(asso.matinput)
}
\arguments{
\item{asso.matinput}{an association matrix estimated using a specified network estimation method.}
}
\value{
A vector network measures of each taxa.
}
\description{
function to compute a network property for each microbial taxa from the estimated association matrix.
}
