% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaNet.R
\name{metaNet}
\alias{metaNet}
\title{metaNet}
\usage{
metaNet(
  feature_abd_list,
  batchvar,
  exposurevar,
  metaData,
  abd_threshold = NULL,
  prev_threshold = NULL,
  covariates = NULL,
  method,
  seed.value
)
}
\arguments{
\item{feature_abd_list}{a list of OTU tables of different studies. Each OTU table is a sample-by-feature data frame.}

\item{batchvar}{name of the batch variable. This variable in data should be a factor variable}

\item{exposurevar}{name of the exposure variable for differential abundance testing.}

\item{metaData}{a data frame of metadata, columns must include exposure, batch as factor, covariates to adjust for.}

\item{abd_threshold}{threshold for abundance of features, cannot be NULL if filter_ft is TRUE.}

\item{prev_threshold}{threshold for prevalence of features, cannot be NULL if filter_ft is TRUE.}

\item{covariates}{optional covariates for batch effect correction - default is NULL.}

\item{method}{method for network estimation - default is sparcc.}

\item{seed.value}{seed for reproducibility}
}
\value{
list of per-feature meta-analytical differential abundance results such as data frame with columns as effect sizes, p-values and q-values, heterogeneity statistics.
}
\description{
Meta-analysis of networks using a pseudo-value approach
}
