% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muffinette.R
\name{muffinette}
\alias{muffinette}
\title{muffinette}
\usage{
muffinette(
  metaAbd,
  batchvar,
  exposurevar,
  metaData,
  filter = TRUE,
  abd_threshold = 0,
  prev_threshold = 0.1,
  topfeatures = NULL,
  batchCorrect = TRUE,
  count,
  net.est.method,
  covariates = NULL,
  ncores = 4,
  analysismethod = "LM",
  verbose = TRUE,
  fixseed = NULL,
  ...
)
}
\arguments{
\item{metaAbd}{a feature-by-sample matrix of abundances across all studies.}

\item{batchvar}{a character vector mapping each observation to its originating study.}

\item{exposurevar}{a character vector mapping each observation to its level of a binary exposure variable.}

\item{metaData}{a data frame of metadata, columns must include exposure, batch as factor, covariates to adjust for, if any.}

\item{filter}{logical. If TRUE, features are filtered according to specified abundance and prevalence thresholds. Default value: TRUE.}

\item{abd_threshold}{numeric; threshold for abundance of features, only used when \code{filter = TRUE}. Default value: 0.}

\item{prev_threshold}{numeric; threshold for prevalence of features, only used when \code{filter = TRUE}. Default value: 0.1.}

\item{topfeatures}{number of features with highest variability to retain for analysis. Default value is \code{NULL} for which all features are kept.}

\item{batchCorrect}{logical. If TRUE, batch effect correction is performed. Default value: TRUE.}

\item{count}{logical. If TRUE, the abundances are count. Default value: FALSE.}

\item{net.est.method}{method for network estimation. Default value: "SparCC".}

\item{covariates}{optional covariates for batch effect correction. Default value: NULL.}

\item{ncores}{number of cores to use for parallelized leave-one-subject-out network estimation.}

\item{analysismethod}{method of meta-analysis.}

\item{verbose}{logical. If TRUE, print out number of iterations and computational time. Default value: TRUE.}

\item{fixseed}{seed for reproducibility}

\item{...}{additional arguments passed to \code{\link{networkEst}} according to the choice specified for \code{net.est.method}.}
}
\value{
a list with
\item{metafits}{data frame with columns as feature names, effect sizes, p-values and q-values, heterogeneity statistics.}
\item{pseudoValues}{sample-by-feature matrix of pseudovalues.}
}
\description{
Meta-analysis of networks using a pseudo-value approach
}
