% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkEst.R
\name{networkEst}
\alias{networkEst}
\title{networkEst}
\usage{
networkEst(x, comp = TRUE, estimethod = "SparCC", ...)
}
\arguments{
\item{x}{abundance matrix}

\item{comp}{logical. If FALSE, the abundances are non-compositional. Default value: TRUE.}

\item{estimethod}{network estimation method. Options include "SparCC" and "SpiecEasi". Default value: SparCC.}

\item{...}{additional arguments appropriate for the network estimation method chosen.}
}
\value{
an association matrix.
}
\description{
function to estimate network based on feature abundances.
}
