% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoValue.R
\name{pseudoValue}
\alias{pseudoValue}
\title{pseudoValue}
\usage{
pseudoValue(thetahatinput, thetahatdropinput, sizegroup)
}
\arguments{
\item{thetahatinput}{network measure of all subjects.}

\item{thetahatdropinput}{network measure excluding a particular subject.}

\item{sizegroup}{number of observations.}
}
\value{
matrix of jackknife pseudovalues.
}
\description{
function to compute the Jackknife pseudovalues of network measures.
}
