% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoValueJK.R
\name{pseudoValueJK}
\alias{pseudoValueJK}
\title{thetahats}
\usage{
pseudoValueJK(thetahatinput, thetahatdropinput, sizegroup)
}
\arguments{
\item{thetahatinput}{network measure of all subjects.}

\item{thetahatdropinput}{network measure excluding a particular subject.}

\item{sizegroup}{number of observations.}
}
\value{
jackknife pseudovalues.
}
\description{
A function to compute the network centrality (i.e. total connectivity)
of each microbial taxa from the association matrix.
}
