% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sohpie_jk.R
\name{sohpie_jk}
\alias{sohpie_jk}
\title{sohpie_jk}
\usage{
sohpie_jk(OTUdat, groupA, groupB, method, seed)
}
\arguments{
\item{OTUdat}{sample-by-feature abundance table after batch effect correction.}

\item{groupA}{Indices of the subjects in the first category of the exposure variable.}

\item{groupB}{Indices of the subjects in the second category of the exposure variable.}

\item{method}{method of network estimation.}

\item{seed}{seed for reproducibility}
}
\value{
sample-by-feature matrix of standardized jackknife values.
}
\description{
function to compute standardized jackknifes using sparCC.
}
