% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparcc.R
\name{sparcc}
\alias{sparcc}
\title{sparcc wrapper}
\usage{
sparcc(x, imax = 20, kmax = 10, alpha = 0.1, Vmin = 1e-04)
}
\arguments{
\item{x}{count data matrix (OTU table)}

\item{imax}{Number of iterations in the outer loop}

\item{kmax}{max iteration steps for SparCC}

\item{alpha}{threshold for strong correlation}

\item{Vmin}{absolute value of correlations below this threshold are considered zero by the inner SparCC loop.}
}
\value{
This will estimate an association matrix (network) for observed OTU table.
}
\description{
SpiecEasi R package, says in his package that
this is "a reimplementation of SparCC algorithm (Friedman et Alm, PLoS Comp Bio, 2012)."
Installation of SpiecEasi can sometimes generate errors, so I have included Dr. Huaying Fang's sparcc wrapper
as one of the functions in this package for the estimation of co-abundance networks. His code was acquired from
CCLasso (Fang et al, Bioinformatics, 2015), provided in GitHub (https://github.com/huayingfang/CCLasso).
}
