% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_features_abd.R
\name{check_features_abd}
\alias{check_features_abd}
\title{check_features_abd}
\usage{
check_features_abd(feature_abd)
}
\arguments{
\item{feature_abd}{a sample-by-feature data frame.}
}
\value{
error messages if feature abundances are (1) missing, (2) negative, (3) not proportions.
}
\description{
function to make sure that feature abundances are neither missing nor negative.
}
