% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultObjects.R
\docType{data}
\name{control_lm_meta}
\alias{control_lm_meta}
\title{Default control settings for meta-analysis}
\format{
An object of class \code{list} of length 11.
}
\usage{
control_lm_meta
}
\description{
A named list of default parameters used in the meta-analysis step.
}
\keyword{datasets}
