% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measureNetwork.R
\name{measureNetwork}
\alias{measureNetwork}
\title{measureNetwork}
\usage{
measureNetwork(asso.matinput)
}
\arguments{
\item{asso.matinput}{An input is an association matrix that is estimated from
the user-provided OTU data.}
}
\value{
A vector containing network centrality of each taxa.
}
\description{
A function to compute the network centrality (i.e. total connectivity)
of each microbial taxa from the association matrix.
}
