% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaNet.R
\name{metaNet}
\alias{metaNet}
\title{metaNet}
\usage{
metaNet(
  feature_abd_list,
  batchvar,
  exposurevar,
  metaData,
  count,
  net.est.method,
  abd_threshold = NULL,
  prev_threshold = NULL,
  covariates = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{feature_abd_list}{a list of OTU tables of different studies. Each OTU table is a sample-by-feature data frame.}

\item{batchvar}{name of the batch variable. This variable in data should be a factor variable.}

\item{exposurevar}{name of the exposure variable for differential abundance testing.}

\item{metaData}{a data frame of metadata, columns must include exposure, batch as factor, covariates to adjust for.}

\item{count}{logical. If TRUE, the abundances are count. Default value: FALSE.}

\item{net.est.method}{method for network estimation. Default value: "SparCC".}

\item{abd_threshold}{threshold for abundance of features.}

\item{prev_threshold}{threshold for prevalence of features.}

\item{covariates}{optional covariates for batch effect correction. Default value: NULL.}

\item{verbose}{logical. If TRUE, print out number of iterations and computational time. Default value: TRUE.}

\item{...}{additional arguments passed to \code{\link{networkEst}}.}
}
\value{
list of per-feature meta-analytical differential abundance results such as data frame with columns as effect sizes, p-values and q-values, heterogeneity statistics.
}
\description{
Meta-analysis of networks using a pseudo-value approach
}
