% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkEst.R
\name{networkEst}
\alias{networkEst}
\title{networkEst}
\usage{
networkEst(x, count = FALSE, estimethod = "SparCC", ...)
}
\arguments{
\item{x}{abundance matrix}

\item{count}{logical. If TRUE, abundances are count instead of proportions. Default value: FALSE.}

\item{estimethod}{network estimation method. Options include "SparCC" and "SpiecEasi". Default value: SparCC.}

\item{...}{additional arguments appropriate for the network estimation method chosen.}
}
\value{
an association matrix.
}
\description{
function to estimate network based on feature abundances.
}
