% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkEst.R
\name{networkEst}
\alias{networkEst}
\title{networkEst}
\usage{
networkEst(X, count = FALSE, est.method, ...)
}
\arguments{
\item{X}{abundance matrix}

\item{count}{logical. If TRUE, abundances are count instead of proportions. Default value: FALSE.}

\item{est.method}{network estimation method. Options include "SparCC" and "SpiecEasi".}

\item{...}{additional arguments appropriate for the network estimation method chosen.}
}
\value{
an association matrix.
}
\description{
Meta-analysis of networks using a pseudo-value approach
}
